feat(base): Cleaning up
